requirements.txt

einops==0.6.0
decord==0.6.0
en_core_web_sm==3.5.0
fairscale==0.4.4
ftfy==6.1.1
horovod==0.26.1
iopath==0.1.10
jsonlines==3.1.0
matplotlib==3.5.2
numpy==1.21.5
omegaconf==2.3.0
opencv_python_headless==4.5.5.64
opendatasets==0.1.22
pandas==1.4.4
Pillow==9.4.0
plotly==5.9.0
pycocoevalcap==1.2
pycocotools==2.0.6
pytest==7.1.2
python_magic==0.4.27
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
scikit_image==0.19.2
scipy==1.9.1
setuptools==63.4.1
spacy==3.5.0
streamlit==1.17.0
timm==0.4.12
torch==1.10.0
torchvision==0.11.2
tqdm==4.64.1
transformers==4.21.3
webdataset==0.2.31
